+++ groups/bsl/bslstl/bslstl_multiset.t.cpp
@@ -1450,13 +1450,13 @@ void TestDriver<KEY, COMP, ALLOC>::testCase22()
 
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
 
             Obj mY(X);  const Obj& Y = mY;
 
             ASSERTV(LINE, 0 == verifyContainer(Y, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    2 * TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    2 * TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
 
             Obj mZ;  const Obj& Z = mZ;
 
@@ -1464,7 +1464,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase22()
 
             ASSERTV(LINE, 0 == verifyContainer(Z, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    2 * TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    2 * TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
         }
 
         CONT.resetIterators();
@@ -1474,7 +1474,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase22()
             mX.insert(BEGIN, END);
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
         }
 
         CONT.resetIterators();
@@ -1488,7 +1488,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase22()
             }
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
         }
 
         ASSERTV(LINE, da.numBlocksInUse(), 0 == da.numBlocksInUse());
